requirements.txt
tensorflow-gpu>=1.9.0

numpy>=1.14.5